+++ docs/Cassandra/FutureClose.php
@@ -25,6 +25,8 @@ namespace Cassandra;
  */
 final class FutureClose implements Future {
 
+    private function __construct() { }
+
     /**
      * Waits for a given future resource to resolve and throws errors if any.
      *