+++ src/FutureValue.stub.php
@@ -10,6 +10,8 @@ namespace Cassandra {
      * @not-serializable
  * @scylladb-struct php_scylladb_future_value
  */    final class FutureValue implements Future {
+        private function __construct() {}
+
         public function get(int|float|null $timeout = null): Value|string|int|float|bool|null {}
 
         /** @return resource */